requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-Talisman>=1.0.0
Flask-Principal>=0.4.0
cryptography>=41.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0 # If using PostgreSQL
pytest>=7.0.0
requests>=2.31.0 
reportlab>=4.0.0 